// src/reecam.types.ts
export type IPCamOptions = {
  /**
   * The target IP address of the camera
   */
  ip: string;
  /**
   * The camera credentials
   */
  credentials: { user: string; pwd: string };
};